Add --avalanche-flags to try to match the input metrics
The --avalanche-flags-for-adjusted-series option requires you to input
the total number of series to target. So it's not convenient to run
mtypes once on a workload to say "just try to approximate its output for
me".

The new --avalanche-flags will do just that, by using the computed
series total as the input for the adjusted series target.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>